abi: v2: Add darkpool v2 ABI
