+++ abi/Cargo.toml
@@ -4,16 +4,29 @@ version = "0.1.0"
 edition = "2021"
 
 [features]
+v1 = []
+v2 = []
+permit2 = ["alloy/rand"]
 relayer-types = [
     "dep:ark-bn254",
     "dep:ark-ec",
     "dep:ark-ff",
     "dep:jf-primitives",
     "dep:num-bigint",
+]
+v1-relayer-types = [
+    "relayer-types",
     "dep:renegade-circuit-types",
     "dep:renegade-circuits",
     "dep:renegade-common",
     "dep:renegade-constants",
+]
+v2-relayer-types = [
+    "relayer-types",
+    "dep:renegade-circuit-types-v2",
+    "dep:renegade-circuits-v2",
+    "dep:renegade-constants-v2",
+    "dep:renegade-crypto-v2",
 ]
 
 [dependencies]
@@ -30,3 +43,8 @@ renegade-circuits = { package = "circuits", git = "https://github.com/renegade-f
 renegade-circuit-types = { package = "circuit-types", git = "https://github.com/renegade-fi/renegade", optional = true }
 renegade-common = { package = "common", git = "https://github.com/renegade-fi/renegade", optional = true }
 renegade-constants = { package = "constants", git = "https://github.com/renegade-fi/renegade", optional = true }
+
+renegade-circuits-v2 = { package = "circuits", path = "/Users/joeykraut/work/renegade/circuits", optional = true }
+renegade-circuit-types-v2 = { package = "circuit-types", path = "/Users/joeykraut/work/renegade/circuit-types", optional = true }
+renegade-constants-v2 = { package = "constants", path = "/Users/joeykraut/work/renegade/constants", optional = true }
+renegade-crypto-v2 = { package = "renegade-crypto", path = "/Users/joeykraut/work/renegade/renegade-crypto", optional = true }
