+++ abi/src/v2/relayer_types/proof_bundles.rs
@@ -1,0 +1,133 @@
+//! Proof bundle conversion helpers
+
+use super::*;
+use crate::v2::IDarkpoolV2;
+use crate::v2::BN254::G1Point;
+use renegade_circuit_types_v2::{traits::BaseType, PlonkProof};
+use renegade_circuits_v2::zk_circuits::valid_balance_create::ValidBalanceCreateStatement;
+
+use ark_ec::AffineRepr;
+use ark_ff::{BigInteger, PrimeField};
+use jf_primitives::pcs::prelude::Commitment as JfCommitment;
+use renegade_constants_v2::MERKLE_HEIGHT;
+
+// -----------------
+// | Proof Bundles |
+// -----------------
+
+impl IDarkpoolV2::NewBalanceDepositProofBundle {
+    /// Create a new proof bundle from a statement and proof
+    pub fn new(statement: ValidBalanceCreateStatement, proof: PlonkProof) -> Self {
+        let merkle_depth = U256::from(MERKLE_HEIGHT);
+        Self {
+            // We use the default merkle height for now
+            merkleDepth: merkle_depth,
+            statement: statement.into(),
+            proof: proof.into(),
+        }
+    }
+}
+
+// -------------------
+// | Statement Types |
+// -------------------
+
+impl From<ValidBalanceCreateStatement> for IDarkpoolV2::ValidBalanceCreateStatement {
+    fn from(statement: ValidBalanceCreateStatement) -> Self {
+        Self {
+            deposit: statement.deposit.into(),
+            newBalanceCommitment: scalar_to_u256(&statement.balance_commitment),
+            newBalancePublicShares: size_vec(
+                statement
+                    .new_balance_share
+                    .to_scalars()
+                    .into_iter()
+                    .map(|s| scalar_to_u256(&s))
+                    .collect(),
+            ),
+            recoveryId: scalar_to_u256(&statement.recovery_id),
+        }
+    }
+}
+
+// --------------------------
+// | Plonk Type Conversions |
+// --------------------------
+
+impl From<PlonkProof> for IDarkpoolV2::PlonkProof {
+    fn from(proof: PlonkProof) -> Self {
+        let evals = proof.poly_evals;
+        Self {
+            wireComms: size_vec(
+                proof
+                    .wires_poly_comms
+                    .into_iter()
+                    .map(convert_jf_commitment)
+                    .collect(),
+            ),
+            zComm: convert_jf_commitment(proof.prod_perm_poly_comm),
+            quotientComms: size_vec(
+                proof
+                    .split_quot_poly_comms
+                    .into_iter()
+                    .map(convert_jf_commitment)
+                    .collect(),
+            ),
+            wZeta: convert_jf_commitment(proof.opening_proof),
+            wZetaOmega: convert_jf_commitment(proof.shifted_opening_proof),
+            wireEvals: size_vec(evals.wires_evals.into_iter().map(fr_to_u256).collect()),
+            sigmaEvals: size_vec(evals.wire_sigma_evals.into_iter().map(fr_to_u256).collect()),
+            zBar: fr_to_u256(evals.perm_next_eval),
+        }
+    }
+}
+
+// ----------------------
+// | Conversion Helpers |
+// ----------------------
+
+/// Size a vector of values to be a known fixed size
+fn size_vec<const N: usize, T>(vec: Vec<T>) -> [T; N] {
+    let size = vec.len();
+    if size != N {
+        panic!("vector is not the correct size: expected {N}, got {size}");
+    }
+    vec.try_into().map_err(|_| ()).unwrap()
+}
+
+/// Convert a point in the BN254 base field to a Uint256
+fn base_field_to_u256(fq: ark_bn254::Fq) -> U256 {
+    let bytes = fq.into_bigint().to_bytes_be();
+    bytes_to_u256(&bytes)
+}
+
+/// Convert a set of big endian bytes to a Uint256
+///
+/// Handles padding as necessary
+fn bytes_to_u256(bytes: &[u8]) -> U256 {
+    let mut buf = [0u8; 32];
+    buf[..bytes.len()].copy_from_slice(bytes);
+    U256::from_be_bytes(buf)
+}
+
+/// Convert a `Fr` to a `U256`
+fn fr_to_u256(fr: ark_bn254::Fr) -> U256 {
+    let bytes = fr.into_bigint().to_bytes_be();
+    bytes_to_u256(&bytes)
+}
+
+/// Convert a point on the BN254 curve to a `G1Point` in the contract's format
+fn convert_g1_point(point: ark_bn254::G1Affine) -> G1Point {
+    let x = point.x().expect("x is zero");
+    let y = point.y().expect("y is zero");
+
+    G1Point {
+        x: base_field_to_u256(*x),
+        y: base_field_to_u256(*y),
+    }
+}
+
+/// Convert a `JfCommitment` to a `G1Point`
+fn convert_jf_commitment(commitment: JfCommitment<ark_bn254::Bn254>) -> G1Point {
+    convert_g1_point(commitment.0)
+}
